Rename observation measurement model file
